requirements.txt
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
gspread==6.0.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1